+++ src/Typstry.jl
@@ -17,8 +17,8 @@ using .Strings: ContextError, Mode, Typst, TypstString, TypstText, @typst_str, c
 export ContextError, Mode, Typst, TypstString, TypstText, @typst_str, code, markup, math, context, show_typst
 
 include("Commands.jl")
-using .Commands: TypstCommand, TypstError, @typst_cmd, julia_mono, preamble, render
-export TypstCommand, TypstError, @typst_cmd, julia_mono, preamble, render
+using .Commands: TypstCommand, TypstError, @typst_cmd, julia_mono, preamble, render, set_preamble
+export TypstCommand, TypstError, @typst_cmd, julia_mono, preamble, render, set_preamble
 
 """
     compile_workload(examples)